+++ dali/imgcodec/decoders/nvjpeg2k.cc
@@ -15,6 +15,8 @@
 #include <string>
 #include <vector>
 #include "dali/imgcodec/decoders/nvjpeg2k.h"
+#include "dali/imgcodec/decoders/nvjpeg2k/nvjpeg2k_memory.h"
+#include "dali/imgcodec/decoders/nvjpeg/nvjpeg_memory.h"
 #include "dali/imgcodec/decoders/nvjpeg/permute_layout.h"
 #include "dali/core/static_switch.h"
