+++ dali/imgcodec/decoders/nvjpeg2k.h
@@ -19,8 +19,7 @@
 #include <memory>
 #include <vector>
 #include "dali/imgcodec/decoders/decoder_parallel_impl.h"
-#include "dali/imgcodec/decoders/nvjpeg/nvjpeg2k_helper.h"
-#include "dali/imgcodec/decoders/nvjpeg/nvjpeg_memory.h"
+#include "dali/imgcodec/decoders/nvjpeg2k/nvjpeg2k_helper.h"
 #include "dali/core/dev_buffer.h"
 #include "dali/core/cuda_stream_pool.h"
 #include "dali/core/cuda_event.h"
