+++ dali/imgcodec/decoders/nvjpeg2k/CMakeLists.txt
@@ -12,17 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_subdirectory(jpeg)
-add_subdirectory(nvjpeg)
-
-if (BUILD_LIBTIFF)
-    add_subdirectory(libtiff)
-endif ()
-
-if (BUILD_NVJPEG2K)
-    add_subdirectory(nvjpeg2k)
-endif ()
-
 collect_headers(DALI_INST_HDRS PARENT_SCOPE)
 collect_sources(DALI_IMGCODEC_SRCS PARENT_SCOPE)
 collect_test_sources(DALI_IMGCODEC_TEST_SRCS PARENT_SCOPE)
