+++ dali/imgcodec/decoders/nvjpeg2k/nvjpeg2k_helper.h
@@ -12,10 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_IMGCODEC_DECODERS_NVJPEG_NVJPEG2K_HELPER_H_
-#define DALI_IMGCODEC_DECODERS_NVJPEG_NVJPEG2K_HELPER_H_
-
-#if NVJPEG2K_ENABLED
+#ifndef DALI_IMGCODEC_DECODERS_NVJPEG2K_NVJPEG2K_HELPER_H_
+#define DALI_IMGCODEC_DECODERS_NVJPEG2K_NVJPEG2K_HELPER_H_
 
 #include <nvjpeg2k.h>
 #include <string>
@@ -155,6 +153,4 @@ inline void cudaResultCheck<nvjpeg2kStatus_t>(nvjpeg2kStatus_t status, const str
 
 }  // namespace dali
 
-#endif  // NVJPEG2K_ENABLED
-
-#endif  // DALI_IMGCODEC_DECODERS_NVJPEG_NVJPEG2K_HELPER_H_
+#endif  // DALI_IMGCODEC_DECODERS_NVJPEG2K_NVJPEG2K_HELPER_H_
